requirements.txt
RPi.GPIO
Flask
paho-mqtt
